+++ UIModuleProject/Packages/com.kwan.uimodule/Runtime/AssemblyInfo.cs
@@ -1,3 +1,4 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("UIModule.RuntimeTests")]
+[assembly: InternalsVisibleTo("UIModule.EditorTests")]